case_study_1/implementation/etl/new_file_trigger/requirements.txt
kafka-python==2.0.2
psycopg2==2.9.7 